Factor out RedisMock to one place
