+++ ores/score_caches/tests/test_redis.py
@@ -1,11 +1,12 @@
 import pytest
 
-from ..redis import Redis
+from ..redis import Redis as RedisScoreCache
+from ...tests.redis_mock import RedisMock
 
 
 @pytest.fixture
 def cache():
-    return Redis.from_config({
+    return RedisScoreCache.from_config({
         'score_caches': {
             'redis': {
                 'host': 'localhost',
@@ -19,25 +20,9 @@ def test_lookup(cache):
     assert cache.lookup('testwiki', 'damaging', 124) == 0.421
 
 
-class RedisMock():
-    def __init__(self):
-        self.values = {}
-
-    def setex(self, key, ttl, value):
-        self.values[key] = value
-        return True
-
-    def get(self, key):
-        return self.values.get(key)
-
-    def delete(self, key):
-        del self.values[key]
-        return True
-
-
 @pytest.fixture
 def cache_mock():
-    return Redis(RedisMock(), 0, 'foo')
+    return RedisScoreCache(RedisMock(), 0, 'foo')
 
 
 @pytest.mark.redis
