+++ ores/task_tracker/tests/test_redis_task_tracker.py
@@ -1,6 +1,7 @@
 import pytest
 
 from ..redis_task_tracker import RedisTaskTracker
+from ...tests.redis_mock import RedisMock
 
 
 @pytest.fixture
@@ -26,25 +27,10 @@ def test_get_in_progress_task(task_tracker):
 
 @pytest.mark.redis
 def test_release(task_tracker):
-    task_tracker.lock('fooo_release', 'value3')
-    assert task_tracker.release('fooo_release')
-    assert task_tracker.get_in_progress_task('fooo_release') is False
-
-
-class RedisMock():
-    def __init__(self):
-        self.values = {}
-
-    def setex(self, key, ttl, value):
-        self.values[key] = value
-        return True
-
-    def get(self, key):
-        return self.values.get(key)
-
-    def delete(self, key):
-        del self.values[key]
-        return True
+    key = 'fooo_release'
+    task_tracker.lock(key, 'value3')
+    assert task_tracker.release(key)
+    assert task_tracker.get_in_progress_task(key) is False
 
 
 @pytest.fixture
@@ -62,6 +48,7 @@ def test_get_in_progress_task_mock(mock_task_tracker):
 
 
 def test_release_mock(mock_task_tracker):
-    mock_task_tracker.lock('fooo_release', 'value3')
-    assert mock_task_tracker.release('fooo_release')
-    assert mock_task_tracker.get_in_progress_task('fooo_release') is False
+    key = 'fooo_release'
+    mock_task_tracker.lock(key, 'value3')
+    assert mock_task_tracker.release(key)
+    assert mock_task_tracker.get_in_progress_task(key) is False
