+++ ores/tests/redis_mock.py
@@ -1,0 +1,14 @@
+class RedisMock:
+    def __init__(self):
+        self.values = {}
+
+    def setex(self, key, ttl, value):
+        self.values[key] = value
+        return True
+
+    def get(self, key):
+        return self.values.get(key)
+
+    def delete(self, key):
+        del self.values[key]
+        return True
